[S07P12B102-280] feat : 쿼리 수정중

+++ BE/ssuk/src/main/java/com/ssafy/ssuk/pot/controller/PotController.java
@@ -1,6 +1,9 @@
 package com.ssafy.ssuk.pot.controller;
 
 import com.ssafy.ssuk.exception.dto.CustomException;
+import com.ssafy.ssuk.exception.dto.ErrorCode;
+import com.ssafy.ssuk.plant.domain.Garden;
+import com.ssafy.ssuk.plant.service.GardenService;
 import com.ssafy.ssuk.pot.domain.Pot;
 import com.ssafy.ssuk.pot.dto.requset.PotDeleteDto;
 import com.ssafy.ssuk.pot.dto.requset.PotInsertDto;
@@ -9,6 +12,8 @@ import com.ssafy.ssuk.pot.service.PotService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,43 +22,47 @@ import java.util.List;
 @RequestMapping("/pot")
 public class PotController {
     private final PotService potService;
+    private final GardenService gardenService;
     private final PotMapper potMapper;
 
     @Autowired
-    public PotController(PotService potService, PotMapper potMapper) {
+    public PotController(PotService potService, GardenService gardenService, PotMapper potMapper) {
         this.potService = potService;
+        this.gardenService = gardenService;
         this.potMapper = potMapper;
     }
 
     //보유한 화분 전체 조회
     @GetMapping("")
-    public ResponseEntity<?> potLIst(@RequestAttribute Integer user_id) {
-        List<Pot> potList = potService.findByUser_Id(user_id);
+    public ResponseEntity<?> potLIst(@RequestAttribute Integer userId) {
+        List<Pot> potList = potService.findByUser_Id(userId);
+        //List<Garden> result = gardenService.findAllByUserId(userId);
 
         return ResponseEntity.ok(potList);
     }
 
-    //화분 상세 조회
-
     //화분 등록
     @PostMapping("")
-    public ResponseEntity<?> save(@RequestAttribute Integer userId, @RequestBody PotInsertDto potInsertDto) throws CustomException {
+    public ResponseEntity<?> save(@RequestAttribute Integer userId, @RequestBody @Validated PotInsertDto potInsertDto, BindingResult bindingResult) throws CustomException {
+        if (bindingResult.hasErrors())
+            throw new CustomException(ErrorCode.INPUT_EXCEPTION);
+
         Pot pot = potMapper.insertDtoToPot(potInsertDto, userId);
 
         potService.save(pot);
 
-        return ResponseEntity.ok().build();
-        //엔티티 변환
+        return ResponseEntity.ok("등록 완료");
     }
 
     //화분 삭제
     @PutMapping("")
-    public ResponseEntity<?> delete(@RequestAttribute Integer userId, @RequestBody PotDeleteDto potDeleteDto) {
-        //Pot pot = potMapper.deleteDtoToPot(potDeleteDto, userId);
+    public ResponseEntity<?> delete(@RequestAttribute Integer userId, @RequestBody @Validated PotDeleteDto potDeleteDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            throw new CustomException(ErrorCode.INPUT_EXCEPTION);
 
         potService.delete(potDeleteDto.getPotId(), userId);
 
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("해제 완료");
     }
 
 }
